+++ packages/affinidi_tdk_vault_iota/test/share_flow_service_test.dart
@@ -1,5 +1,5 @@
 import 'package:affinidi_tdk_cryptography/affinidi_tdk_cryptography.dart';
-import 'package:affinidi_tdk_vault/affinidi_tdk_vault.dart';
+import 'package:affinidi_tdk_vault_iota/affinidi_tdk_vault_iota.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';